requirements_gcp.txt
ypy-websocket==0.8.4
webencodings==0.5.1
oauthlib==3.2.2
typeguard==4.4.2
cryptography==44.0.2
websockets==15.0.1
attrs==25.3.0
aiosqlite==0.21.0
markdown-it-py==3.0.0
seaborn==0.12.2
mpmath==1.3.0
jupyterlab==3.6.8
nvidia-cusolver-cu12==11.6.0.99
gpustat==1.0.0
webcolors==24.11.1
nbclassic==1.2.0
python-dateutil==2.9.0.post0
google-cloud-artifact-registry==1.15.2
greenlet==3.1.1
colorful==0.5.6
wrapt==1.17.2
nest-asyncio==1.6.0
python-json-logger==3.3.0
blessed==1.20.0
nvidia-nvtx-cu12==12.4.99
gcsfs==2025.3.2
distlib==0.3.9
notebook==6.5.7
pyarrow-hotfix==0.6
scipy==1.11.4
nvidia-nvjitlink-cu12==12.4.99
keyring==25.6.0
argon2-cffi==23.1.0
jeepney==0.9.0
aiohttp-cors==0.8.1
simsimd==6.2.1
google-cloud-monitoring==2.27.1
executing==2.2.0
beautifulsoup4==4.13.3
asttokens==3.0.0
zipp==3.21.0
rsa==4.9
platformdirs==4.3.7
joblib==1.4.2
jupyter_server==2.15.0
GitPython==3.1.44
pins==0.8.7
pycparser==2.22
nvidia-nccl-cu12==2.20.5
kfp-server-api==2.0.5
types-python-dateutil==2.9.0.20241206
aiofiles==22.1.0
nbdime==3.2.0
google-cloud-resource-manager==1.14.2
uvloop==0.21.0
ipython==8.21.0
matplotlib==3.7.3
plotly==6.0.1
traitlets==5.14.3
grpcio-status==1.49.0rc1
ptyprocess==0.7.0
opentelemetry-proto==1.27.0
nbformat==5.10.4
nvidia-cuda-nvrtc-cu12==12.4.99
ipython-genutils==0.2.0
pillow==11.1.0
mistune==3.1.3
nvidia-curand-cu12==10.3.5.119
prettytable==3.16.0
albucore==0.0.23
argon2-cffi-bindings==21.2.0
pyzmq==26.4.0
jupyter-ydoc==0.2.5
gitdb==4.0.12
Jinja2==3.1.6
google-auth==2.38.0
sqlparse==0.5.3
virtualenv==20.30.0
urllib3==1.26.20
opentelemetry-exporter-otlp-proto-http==1.27.0
opencv-python-headless==4.11.0.86
pydantic==1.10.21
phik==0.12.4
idna==3.10
sniffio==1.3.1
uvicorn==0.34.0
packaging==24.2
kfp-pipeline-spec==0.2.2
nvidia-cuda-runtime-cu12==12.4.99
pydata-google-auth==1.9.1
filelock==3.18.0
google-auth-httplib2==0.2.0
h11==0.14.0
SecretStorage==3.3.3
proto-plus==1.26.1
importlib_resources==6.5.2
jupyter_server_terminals==0.5.3
msgpack==1.1.0
lightning==2.5.1.post0
stack-data==0.6.3
google-api-python-client==1.8.0
simpervisor==1.0.0
pyogrio==0.10.0
pyarrow==14.0.2
overrides==7.7.0
pandocfilters==1.5.1
jsonschema==4.23.0
pandas==2.0.3
parsy==2.1
jaraco.functools==4.1.0
bleach==6.2.0
shapely==2.1.0
google-cloud-functions==1.20.2
pure_eval==0.2.3
google-cloud-aiplatform==1.88.0
google-cloud-datastore==1.15.5
pexpect==4.9.0
numpy==1.25.2
tomli==2.2.1
grpcio==1.72.0rc1
tokenizers==0.21.1
jedi==0.19.2
prompt_toolkit==3.0.50
fastrlock==0.8.3
SQLAlchemy==2.0.40
jupyterlab_widgets==3.0.14
watchfiles==1.0.5
aiosignal==1.3.2
cloudpickle==3.1.1
toolz==0.12.1
opentelemetry-exporter-otlp==1.27.0
jupyter_core==5.7.2
huggingface-hub==0.30.2
docstring_parser==0.16
multimethod==1.12
fastjsonschema==2.21.1
cffi==1.17.1
nvidia-cublas-cu12==12.4.2.65
pip==25.0.1
fsspec==2025.3.2
jsonpointer==3.0.0
isoduration==20.11.0
requests-toolbelt==0.10.1
ray==2.44.1
fonttools==4.57.0
websocket-client==1.8.0
google-api-core==1.34.1
json5==0.12.0
sqlglot==19.9.0
humanize==4.12.2
Deprecated==1.2.18
atpublic==4.1.0
MarkupSafe==3.0.2
tinycss2==1.4.0
jsonschema-specifications==2024.10.1
torch-xla==2.4.0
albumentations==2.0.5
ansicolors==1.1.8
papermill==2.6.0
dacite==1.9.2
htmlmin==0.1.12
soupsieve==2.6
Pygments==2.19.1
bigframes==0.22.0
click==8.1.8
colorama==0.4.6
wandb==0.19.10
parso==0.8.4
opentelemetry-semantic-conventions==0.48b0
lightning-utilities==0.14.3
opencensus==0.11.4
certifi==2025.1.31
PyJWT==2.10.1
httplib2==0.22.0
protobuf==3.20.3
py-spy==0.4.0
typing_extensions==4.13.2
dataproc_jupyter_plugin==0.1.80
triton==3.0.0
six==1.17.0
llvmlite==0.41.1
mdit-py-plugins==0.4.2
opentelemetry-exporter-otlp-proto-grpc==1.27.0
sympy==1.13.3
cachetools==5.5.2
y-py==0.6.2
narwhals==1.34.1
jupyter_server_fileid==0.9.3
uritemplate==3.0.1
propcache==0.3.1
PyYAML==6.0.2
google-resumable-media==2.7.2
ipykernel==6.29.5
tangled-up-in-unicode==0.2.0
kiwisolver==1.4.8
terminado==0.18.1
pytorch-lightning==2.5.1.post0
googleapis-common-protos==1.69.2
backports.tarfile==1.2.0
keyrings.google-artifactregistry-auth==1.1.2
psutil==5.9.3
google-cloud-jupyter-config==0.0.10
defusedxml==0.7.1
multidict==6.4.3
pyasn1_modules==0.4.2
db-dtypes==1.4.2
mdurl==0.1.2
pyOpenSSL==25.0.0
visions==0.7.5
scikit-learn==1.6.1
tqdm==4.67.1
Send2Trash==1.8.3
entrypoints==0.4
timm==1.0.15
matplotlib-inline==0.1.7
pyasn1==0.6.1
ydata-profiling==4.6.0
aiohttp==3.9.5
wcwidth==0.2.13
pyproj==3.7.1
jupyter_server_proxy==4.4.0
jupyterlab_git==0.44.0
nbconvert==7.16.6
nbclient==0.10.2
ipywidgets==8.1.6
threadpoolctl==3.6.0
contourpy==1.3.1
async-timeout==4.0.3
docker==7.1.0
tabulate==0.9.0
smart-open==7.1.0
jupyterlab_pygments==0.3.0
PyWavelets==1.8.0
fastapi==0.115.12
notebook_shim==0.2.4
google-cloud-language==2.17.1
google-cloud-iam==2.18.3
pyparsing==3.2.3
jupyterlab_server==2.27.3
fqdn==1.5.1
arrow==1.3.0
jupyter-events==0.12.0
tenacity==9.1.2
grpc-google-iam-v1==0.14.2
safetensors==0.5.3
jaraco.context==6.0.1
torchmetrics==1.7.1
ibis-framework==7.1.0
tornado==6.4.2
babel==2.17.0
jupyter_server_ydoc==0.8.0
jupyter-server-mathjax==0.2.6
oauth2client==4.1.3
google-cloud-bigquery-connection==1.18.2
annotated-types==0.7.0
absl-py==2.2.2
jupyter-http-over-ws==0.0.8
setuptools==65.5.0
torchvision==0.19.0+cu124
networkx==3.4.2
linkify-it-py==2.0.3
nvidia-ml-py==11.495.46
rfc3339-validator==0.1.4
appdirs==1.4.4
memray==1.17.1
patsy==1.0.1
jupyter_client==7.4.9
multipledispatch==1.0.0
pluggy==1.5.0
frozenlist==1.5.0
rfc3986-validator==0.1.1
exceptiongroup==1.2.2
decorator==5.2.1
geopandas==1.0.1
wordcloud==1.9.4
rich==13.9.4
jupytext==1.17.0
opencensus-context==0.1.3
referencing==0.36.2
beatrix_jupyterlab==2024.66.154055
debugpy==1.8.14
requests-oauthlib==2.0.0
charset-normalizer==3.4.1
textual==3.0.1
comm==0.2.2
dm-tree==0.1.9
opentelemetry-api==1.27.0
statsmodels==0.14.4
docker-pycreds==0.4.0
httptools==0.6.4
ipython-sql==0.5.0
nvidia-cuda-cupti-cu12==12.4.99
widgetsnbextension==4.0.14
opentelemetry-exporter-otlp-proto-common==1.27.0
bidict==0.23.1
cloud-tpu-client==0.10
rpds-py==0.24.0
transformers==4.51.3
anyio==4.9.0
pydantic==2.11.3
google-cloud-storage==2.14.0
stringzilla==3.12.5
lz4==4.4.4
setproctitle==1.3.5
smmap==5.0.2
uc-micro-py==1.0.3
regex==2024.11.6
jaraco.classes==3.4.0
typing-inspection==0.4.0
yarl==1.19.0
gymnasium==1.0.0
google-crc32c==1.7.1
Farama-Notifications==0.0.4
pytz==2025.2
sentry-sdk==2.27.0
google-auth-oauthlib==1.2.1
uri-template==1.3.0
google-cloud-bigquery-storage==2.30.0
time-machine==2.16.0
cupy-cuda12x==13.4.1
Cython==3.0.12
nvidia-cufft-cu12==11.2.0.44
kernels-mixer==0.0.15
tzdata==2025.2
requests==2.32.3
notebook_executor==0.2
google-cloud-core==2.4.3
cycler==0.12.1
ormsgpack==1.7.0
starlette==0.46.1
importlib_metadata==8.4.0
kubernetes==26.1.0
pydantic_core==2.33.1
more-itertools==10.6.0
kfp==2.5.0
nvidia-cusparse-cu12==12.3.0.142
retrying==1.3.4
tensorboardX==2.6.2.2
nvidia-cudnn-cu12==9.1.0.70
prometheus_client==0.21.1
ImageHash==4.3.1
xxhash==3.5.0
torch==2.4.0+cu124
python-dotenv==1.1.0
opentelemetry-sdk==1.27.0
google-cloud-bigquery==3.25.0
numba==0.58.1
pendulum==3.0.0